jupyter/requirements.txt
python_dotenv
rdflib
oxrdflib
docarray
langchain
langchain-community
langchain[docarray]
langchain-ollama
langgraph
langgraph-checkpoint-sqlite
pydantic
neo4j
langchain_aws
#lama index dependencies
llama-index
llama-index-llms-bedrock
llama-index-llms-anthropic
llama-index-llms-ollama
llama-index-graph-stores-neo4j
llama-index-readers-graphdb-cypher